Add files via upload
This project is a work in progress. The ultimate project will leverage a regressor from scikit-learn and will take the inputs and produce a most accurate estimate for what a client can expect to pay for a job. 
A current state of the app can be seen at: 
http://silverbackgino.herokuapp.com


app1.py - source code

Libraries - requirements.txt